refactor: reestruturação da lógica e organização do código
