+++ src/microcontroller/infraredSensor/infrared.h
@@ -5,12 +5,10 @@
 
 typedef struct {
     uint8_t pin;
-    uint8_t side;
     int edgeDetection;
 } ModuleInfrared;
 
-void infraredInit(ModuleInfrared *infrared, uint8_t pin, int edgeDetection, uint8_t side);
+void infraredInit(ModuleInfrared *infrared, uint8_t pin, int edgeDetection);
 int verifyEdge(ModuleInfrared *infrared);
-void debugInfraredState(ModuleInfrared *infrared);
 
 #endif
