+++ src/microcontroller/motor/motor.h
@@ -5,8 +5,7 @@
 #include <stdbool.h>
 
 typedef struct {
-    uint8_t rightMotorEnable;
-    uint8_t leftMotorEnable;
+    uint8_t motorEnable;
     uint8_t rightMotorPWMF;
     uint8_t leftMotorPWMF;
     uint8_t rightMotorPWMB;
@@ -14,8 +13,8 @@ typedef struct {
     float tension;
 } ModuleMotor;
 
-void motorInit(ModuleMotor *motor, uint8_t rightMotorEnable, uint8_t leftMotorEnable, 
-uint8_t rightMotorPWMF, uint8_t leftMotorPWMF, uint8_t rightMotorPWMB, uint8_t leftMotorPWMB, float tension);
+void motorInit(ModuleMotor *motor, uint8_t motorEnable, uint8_t rightMotorPWMF,
+ uint8_t leftMotorPWMF, uint8_t rightMotorPWMB, uint8_t leftMotorPWMB, float tension);
 void moveForward(ModuleMotor *motor);
 void moveBackward(ModuleMotor *motor);
 void turnLeft(ModuleMotor *motor);
